+++ react/components/ProductSpecification/ProductSpecification.schema.ts
@@ -1,0 +1,12 @@
+const ProductSpecificationSchema = {
+  title: 'Especificação de produto',
+  type: 'object',
+  properties: {
+    specificationName: {
+      title: 'Nome da especificação',
+      type: 'string',
+    },
+  },
+};
+
+export { ProductSpecificationSchema };