+++ react/components/ProductSpecification/ProductSpecification.types.ts
@@ -1,0 +1,5 @@
+interface ProductSpecificationProps {
+  specificationName: string;
+}
+
+export { ProductSpecificationProps };