+++ react/hooks/useProductSpecification/index.ts
@@ -1,0 +1,3 @@
+import useProductSpecification from './useProductSpecification';
+
+export default useProductSpecification;