+++ react/useProductSpecification.ts
@@ -1,0 +1,3 @@
+import useProductSpecification from './hooks/useProductSpecification';
+
+export default useProductSpecification;