+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added new **useProductSpecification** hook.
+- Added new **ProductSpecification** component.
+- Added new **useCheckout** hook.
+
 ## [1.3.0] - 20-05-2022
 
 ### Added